Migrate EditManualMovementSpeedsWindow -> DialogPage
- Simplify naming
- MatterHackers/MCCentral#3752
Migrate to DialogPage
